ulogd: logging of PHYSIN / PHYSOUT
The use of bridge interfaces is quite common on routers or when
virtualization (lxc/kvm) is in use. Sometimes there are quite a lot of
interfaces on a bridge. The PHYSIN and PHYSOUT information is quite
important on those systems. Important enough to not use ulogd with
NFLOG.

This patch fixes this for every output that uses the printpkt function.